Change controller concern to model concern
- Move the logic of `number_with_country_code` into concern
- Delete helper module
- Define shared_examples for `phone_number_localization` to test both `PatientPhoneNumber` and `PhoneNumberAuthentication`
