+++ app/models/phone_number_authentication.rb
@@ -1,6 +1,6 @@
 class PhoneNumberAuthentication < ApplicationRecord
   include PgSearch::Model
-  include PhoneNumberLocalization
+  include PhoneNumberLocalizable
 
   USER_AUTH_MAX_FAILED_ATTEMPTS = Integer(ENV["USER_AUTH_MAX_FAILED_ATTEMPTS"] || 5).freeze
   USER_AUTH_LOCKOUT_IN_MINUTES = Integer(ENV["USER_AUTH_LOCKOUT_IN_MINUTES"] || 20).freeze
