+++ spec/models/patient_phone_number_spec.rb
@@ -1,4 +1,5 @@
 require "rails_helper"
+require Rails.root.join "spec/models/concerns/phone_number_localizable_spec.rb"
 
 RSpec.describe PatientPhoneNumber, type: :model do
   describe "Associations" do
@@ -102,5 +103,9 @@ RSpec.describe PatientPhoneNumber, type: :model do
       }.to change(ExotelPhoneNumberDetail.where(patient_phone_number: patient_phone_number), :count)
         .from(0).to(1)
     end
+  end
+
+  describe "localized_phone_number" do
+    it_behaves_like "phone_number_localizable"
   end
 end
